+++ rector.php
@@ -27,4 +27,5 @@ return RectorConfig::configure()
     ])
     ->withPaths([
         __DIR__ . '/src',
+        __DIR__ . '/tests',
     ]);