+++ src/models/Tenant.php
@@ -152,9 +152,9 @@ class Tenant extends ActiveRecord implements DraftStatusAttributeInterface
     public function validateCookieDomain(): void
     {
         if (
-            str_starts_with((string) $this->cookie_domain, 'http')
-            || !str_contains($this->url, ltrim((string) $this->cookie_domain, '.'))
-            || !preg_match('/^[a-z.]/', (string) $this->cookie_domain)
+            str_starts_with((string)$this->cookie_domain, 'http')
+            || !str_contains($this->url, ltrim((string)$this->cookie_domain, '.'))
+            || !preg_match('/^[a-z.]/', (string)$this->cookie_domain)
         ) {
             $this->addInvalidAttributeError('cookie_domain');
         }
@@ -188,6 +188,12 @@ class Tenant extends ActiveRecord implements DraftStatusAttributeInterface
         }
 
         return true;
+    }
+
+    public function afterDelete(): void
+    {
+        TenantCollection::invalidateCache();
+        parent::afterDelete();
     }
 
     public function isDeletable(): bool