+++ src/web/UrlManager.php
@@ -33,12 +33,6 @@ class UrlManager extends \davidhirtz\yii2\skeleton\web\UrlManager
         return Url::ensureScheme($url, $scheme);
     }
 
-    public function createDraftUrl(array|string $params): string
-    {
-        $tenant = $this->getTenantFromParams($params, true);
-        return Url::draft($tenant->getHostInfo()) . $this->createUrl($params);
-    }
-
     public function createUrl($params): string
     {
         $tenant = $this->getTenantFromParams($params, true);
@@ -58,8 +52,6 @@ class UrlManager extends \davidhirtz\yii2\skeleton\web\UrlManager
         $tenant = $this->getTenantFromRequest($request);
         $this->setTenant($tenant);
 
-        $this->defaultLanguage = $tenant->language;
-
         if ($tenant->isDraft()) {
             Yii::$app->getResponse()->getHeaders()->set('X-Robots-Tag', 'none');
         }
@@ -67,9 +59,17 @@ class UrlManager extends \davidhirtz\yii2\skeleton\web\UrlManager
         return parent::parseRequest($request);
     }
 
-    protected function setTenant(Tenant $tenant): void
+    public function setTenant(Tenant $tenant): void
     {
         Yii::$app->set('tenant', $tenant);
+        Yii::$app->language = $tenant->language;
+
+        $this->setHostInfo($tenant->getHostInfo());
+    }
+
+    protected function setApplicationLanguage(Request $request): void
+    {
+        Yii::$app->language = Yii::$app->get('tenant')->language;
     }
 
     protected function setCookieDomain(string $domain): void
@@ -96,7 +96,7 @@ class UrlManager extends \davidhirtz\yii2\skeleton\web\UrlManager
             return $tenant;
         }
 
-        $tenant = current(TenantCollection::getAll());
+        $tenant = TenantCollection::getDefault();
 
         Yii::debug("Tenant not found by host name or path info, using tenant: $tenant->name", __METHOD__);
         return $tenant;